Disitool is a small Python program to manipulate embedded digital signatures.
* delete a signature: disitool.py delete signed-file unsigned-file
* copy a signature: disitool.py copy signed-source-file unsigned-file signed-file
* extract a signature: disitool.py extract signed-file signature
* add a signature: disitool.py add signature unsigned-file signed-file
* inject data after the authenticode signature: disitool.py inject
  [--paddata] signed-source-file data-file signed-destination-file

It is not a tool to digitally sign executables, use signtool for
this. When you add or copy a signature from one file to another
file, the signature will not be valid.

disitool uses pefile, you’ll need to install this first. This new
version (V0.2) will update the PE header checksum.

http://blog.didierstevens.com/programs/disitool/